==> processing/04-merge.ipynb <==
# Merge

Combine data files into a CSV that's ready for analysis

import pandas as pd

Import data files

deaths_df = pd.read_csv(
    "../input/processed/death-records.csv",
    parse_dates=["date_of_death", "date_of_birth"],
    dtype={
        "last_name": str,
        "first_name": str,
        "middle_name": str,
        "sex": str,
        "place_of_birth_state_or_foreign_country": str,
        "place_of_death_county": str,
        "fathers_last_name": str,
        "file_name": str,
        "file_year": int,
        "year_of_death": int,
        "death_facility_name_location": str,
        "place_of_death_address_street_number": str,
        "place_of_death_address_street": str,
        "place_of_death_city": str,
        "father's_last_name": str,
        "place_of_death_facility": str,
        "place_of_death_facility_name_location": str,
        "place_of_death_address_street_name": str,
        "export_date": str,
        "father_last_name": str
    }
)

pops_df = pd.read_csv("../input/raw/dof/county-populations.csv")

recent_temps_df = pd.read_csv(
    "../input/processed/recent-temps-by-county.csv",
    parse_dates=["date"]
)

historic_temps_df = pd.read_csv(
    "../input/processed/historic-temps-by-county.csv",
)

Aggregate

deaths_by_day = deaths_df.groupby([
    "date_of_death",
    "place_of_death_county"
]).size().rename("deaths").reset_index().rename(
    columns={"date_of_death":"date", "place_of_death_county":"county"}
)

Backfill

backfilled_deaths_by_day = (
    deaths_by_day.set_index(["county", "date"])
    .unstack(["county"])
    .stack("county", dropna=False)
    .reset_index()
    .sort_values(["county", "date"])
).fillna(0)

Annotate

backfilled_deaths_by_day['day_name'] = backfilled_deaths_by_day['date'].dt.day_name()

backfilled_deaths_by_day['is_weekend'] = backfilled_deaths_by_day.day_name.isin(
    ['Saturday', 'Sunday']
)

backfilled_deaths_by_day['day'] = backfilled_deaths_by_day['date'].dt.day

backfilled_deaths_by_day['month'] = backfilled_deaths_by_day['date'].dt.month

backfilled_deaths_by_day['year'] = backfilled_deaths_by_day['date'].dt.year

Merge pops to deaths

pops_by_year = pops_df.set_index("county").stack().reset_index().rename(columns={
    "level_1": "year",
    0: "population"
})

pops_by_year.year = pops_by_year.year.astype(int)

pops_by_year.county = pops_by_year.county.str.upper()

pops_merge = backfilled_deaths_by_day.merge(pops_by_year, on=["county", "year"], how="inner")

len(pops_merge)

assert len(pops_merge) == len(backfilled_deaths_by_day)

Merge temps

recent_temps_df.county = recent_temps_df.county.str.upper()

recent_temps_merge = pops_merge.merge(
    recent_temps_df,
    on=["county", "date"],
    how="inner"
)

#assert len(recent_temps_merge) == len(pops_merge) == len(recent_temps_df)

recent_temps_merge.head()

historic_temps_df.county = historic_temps_df.county.str.upper()

historic_temps_merge = recent_temps_merge.merge(
    historic_temps_df,
    on=["county", "month", "day"],
    how="inner"
).rename(columns={'tmax_95':'heat_event_threshold_f'})

assert (historic_temps_merge.month.unique() == [5, 6, 7, 8, 9, 10]).all()

historic_temps_merge['heat_event_threshold_f'] = historic_temps_merge['heat_event_threshold_f'].round()

Calculate

historic_temps_merge['is_heat_event'] = historic_temps_merge.tmax_f >= historic_temps_merge.heat_event_threshold_f

Export

historic_temps_merge.to_csv("../output/totals-by-day.csv", index=False)
